+++ docs/reference/changelog.md
@@ -46,6 +46,7 @@ Unreleased changes
 ### Internal
 
 - Simplify and refactor `interpolate_color` to not depend on `px.colors.find_intermediate_color` ({gh-pr}`253`)
+- Improve type narrowing using `typing.TypeIs` ({gh-pr}`???`)
 
 ---